Imports the sound
This inputs the sound so that there is a click before the reaction time is tested. This was difficult to figure out and took me about an hour to learn the various ways to add the sound. I need to find out how to get it to play automatically now.